requirements.txt
asgiref==3.11.0
Django==4.2.7
django-cors-headers==4.9.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
pillow==12.0.0
PyJWT==2.10.1
pytz==2025.2
sqlparse==0.5.3
